requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
airportsdata==20241001
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
autoflake==2.3.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
datasets==3.1.0
dill==0.3.8
diskcache==5.6.3
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
huggingface-hub==0.26.3
idna==3.10
iniconfig==2.0.0
interegular==0.3.3
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark==1.2.2
llm_structured_output==0.0.19
MarkupSafe==3.0.2
mlx==0.21.0
mlx-lm==0.20.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy==1.13.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
outlines==0.1.7
outlines_core==0.1.17
packaging==24.2
# Editable Git install with no remote (pal==0.1)
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.0
psutil==6.1.0
pyarrow==18.1.0
pycountry==24.6.1
pydantic==2.10.2
pydantic_core==2.27.1
pyflakes==3.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.21.0
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.20.3
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
xxhash==3.5.0
yarl==1.18.3
